quantity feature added

+++ pantryapp/app/page.js
@@ -3,7 +3,7 @@ import { Typography, Button , Modal, TextField} from '@mui/material';
 import Box from '@mui/material/Box';
 import Stack from '@mui/material/Stack';
 import { Firestore } from 'firebase/firestore';
-import { addDoc,collection,getDocs,query,doc , setDoc, docRef, deleteDoc} from 'firebase/firestore';
+import { addDoc,collection,getDocs,query,doc , setDoc, docRef, deleteDoc, getDoc} from 'firebase/firestore';
 import { useEffect, useState } from 'react';
 import {firestore} from '../firebase';
 export default function Home() {
@@ -17,7 +17,7 @@ export default function Home() {
     const docs= await getDocs(snapshot)
     const pantryList=[]
     docs.forEach((doc) =>{
-      pantryList.push({name: doc.id, count:doc.data()})
+      pantryList.push({name: doc.id, ...doc.data()})
     })
     console.log(pantryList)
     setPantry(pantryList)
@@ -28,13 +28,27 @@ export default function Home() {
   const addItem=async(item) => {
     const docRef = doc(collection(firestore, 'pantry'), item)
     //Check if it exists
-    
-    await setDoc(docRef,{})
+    const docSnap= await getDoc(docRef)
+    if (docSnap.exists()) {
+      const {count} = docSnap.data()
+      await setDoc(docRef,{count: count+1})
+    }
+    else{
+      await setDoc(docRef,{count: 1})
+    }
     await updatePantry()
   }
   const removeItem=async(item) => {
     const docRef = doc(collection(firestore, 'pantry'), item)
-    await deleteDoc(docRef)
+    const docSnap= await getDoc(docRef)
+    if (docSnap.exists()) {
+      const {count} = docSnap.data()
+      if(count==1){
+        await deleteDoc(docRef)
+      }
+    else{
+      await setDoc(docRef,{count: count-1})
+    }}
     await updatePantry()
   }
   return (
@@ -91,7 +105,7 @@ export default function Home() {
           </Stack>
         </Box>
       </Modal> 
-    <Button variant="contained" onClick={handleOpen}>Add</Button>
+    <Button variant="contained" color='success' onClick={handleOpen}>Add</Button>
     <Box border={'1px solid #333'}>
     <Box width="800px" height="100px" bgcolor={'#ADD8E6'} display={'flex'} justifyContent={'center'} alignItems={'center'}>
     <Typography variant={'h2'} color={'#333'} textAlign={'center'}>
@@ -99,10 +113,10 @@ export default function Home() {
     </Typography>
     </Box>
     <Stack width="800px" height="300px" spacing={2} overflow={'auto'} >
-      {pantry.map((i)=>(
+      {pantry.map((item)=>(
         
         <Box 
-        key={i}
+        key={item.name}
         width="100%"
         minHeight="150px"
         display={'flex'}
@@ -113,10 +127,13 @@ export default function Home() {
         >
           <Typography variant={'h3'} color={'#333'} textAlign={'center'}>
             {
-              i.charAt(0).toUpperCase() + i.slice(1)
+              item.name.charAt(0).toUpperCase() + item.name.slice(1)
             }
           </Typography>
-        <Button variant='contained' onClick={()=> removeItem(i)}>Remove</Button>
+          <Typography variant={'h3'} color={'#333'} textAlign={'center'}>
+            Quantity: {item.count}
+          </Typography>
+        <Button variant='contained' color='error' onClick={()=> removeItem(item.name)}>Remove</Button>
         </Box>
         //</Stack>
       ))}
